refactor(utils): rename `integer` utility as `number` utility
