+++ meson.build
@@ -169,10 +169,10 @@ sources = [
   'src/utils/file.c',
   'src/utils/hashset.c',
   'src/utils/hashtable.c',
-  'src/utils/integer.c',
   'src/utils/linkedlist.c',
   'src/utils/logging.c',
   'src/utils/notifier.c',
+  'src/utils/number.c',
   'src/utils/queue.c',
   'src/utils/string.c',
   'src/utils/tqueue.c',
